Add arbiter for handling server startup
